Rwa 132 (#21)
* move header to the left

* rename header and fix sign out button

* add more options to the hearder

* add task-list router

* add task-list template

* add radio buttons and js to show/hide one or the other view

* migrate myTasks view to the new merged view

* display available tasks when radio button is checked

* redirect to task-list when claim and unclaim actions executed

* fix unit test

* add filter by location component to the task-list view

* remove old task list screen

* remove old task list code

* refactor: move routers and controllers to its own files

* make due date bold (#22)

* make due date bold

* fix caseRef content

* update umclaim task to release task

* add one more action to the details

* add go to case action

* fix action button  styling

* refactor: rename button
