+++ src/main/assets/scss/header.scss
@@ -1,9 +1,13 @@
-.sign-out-button {
+#sign-out-button {
   position: absolute;
   right: 10px;
   top: 5px;
-  color: #767676;
   float: right;
   padding: 10px;
   z-index: 100;
+  
+}
+
+.govuk-header__logo {
+  width: 24.33%;
 }
