+++ src/main/assets/scss/main.scss
@@ -4,7 +4,7 @@
 @import 'next-step';
 @import 'summary-list';
 @import 'tabs';
-@import 'sign-out';
-@import 'my-case-work';
+@import 'header';
+@import 'task-list';
 @import 'filter-by-location';
