--- src/main/assets/scss/my-case-work.scss
@@ -1,22 +1,0 @@
-.my-task-header {
-  border-bottom: none;
-  padding-bottom: 30px;
-}
-
-.my-tasks-table {
-  white-space: nowrap;
-}
-
-.wa-action-task {
-  display: none;
-  padding: 5px 0px 20px 0;
-}
-
-.govuk-details__text {
-  border-left: none;
-  padding: 15px 15px 0px 15px;
-}
-
-.filter-by-location-details {
-  padding-left: 15px;
-}
