+++ src/main/models/myCaseWorkModel.ts
@@ -1,15 +1,18 @@
 import { Task } from '../models/task';
 import { Location } from '../models/location';
 
-export class MyCaseWorkModel {
+const task0 = new Task('1549-6338-2756-6773', 'Lala Joji', 'Human', 'Taylor House', 'Review respondent evidence', 'Today', 3, 'today');
+const task1 = new Task('1549-2353-5684-8902', 'Rachel Smith', 'Human', 'Birmingham', 'Review respondent evidence', 'Today', 3, 'today');
+const task2 = new Task('1549-9061-9513-0004', 'Nonasky Pecki', 'EEA', 'Taylor House', 'Create case summary', '04 Dec', 2, 'future');
+const task3 = new Task('1549-9083-5480-6960', 'Wauanda Jik', 'Human Rights', 'Taylor House', 'Review appellant case', '+2 days', 4, 'past');
+const task4 = new Task('1549-4765-3206-5586', 'Kili Muso', 'Protection', 'Taylor House', 'Review respondent evidence', 'Today', 3, 'today');
+const task5 = new Task('1549-5366-1108-0150', 'Mankay Lit', 'Revocation', 'Taylor House', 'Review appellant case', '14 Dec', 1, 'future');
+const task6 = new Task('1549-2345-7854-9786', 'Wina Palic', 'Human Rights', 'Taylor House', 'Review respondent evidence', '+4 days', 5, 'past');
+
+export class MyCaseWorkModel {  
 
   static getMyAvailableTasks(): Array<Task> {
     const tasks: Array<Task> = [];
-
-    const task0 = new Task('1549 6338 2756 6773', 'Lala Joji', 'Human', 'Taylor House', 'Review respondent evidence', 'Today', 3, 'today');
-    const task1 = new Task('1549 2353 5684 8902', 'Rachel Smith', 'Human', 'Birmingham', 'Review respondent evidence', 'Today', 3, 'today');
-    const task2 = new Task('1549 9061 9513 0004', 'Nonasky Pecki', 'EEA', 'Taylor House', 'Create case summary', '04 Dec', 2, 'future');
-    const task3 = new Task('1549 9083 5480 6960', 'Wauanda Jik', 'Human Rights', 'Taylor House', 'Review appellant case', '+2 days', 4, 'past');
 
     tasks.push(task0);
     tasks.push(task1);
@@ -22,10 +25,6 @@ export class MyCaseWorkModel {
   static getMyFilteredAvailableTasks(): Array<Task> {
     const tasks: Array<Task> = [];
 
-    const task0 = new Task('1549 6338 2756 6773', 'Lala Joji', 'Human', 'Taylor House', 'Review respondent evidence', 'Today', 3, 'today');
-    const task2 = new Task('1549 9061 9513 0004', 'Nonasky Pecki', 'EEA', 'Taylor House', 'Create case summary', '04 Dec', 2, 'future');
-    const task3 = new Task('1549 9083 5480 6960', 'Wauanda Jik', 'Human Rights', 'Taylor House', 'Review appellant case', '+2 days', 4, 'past');
-
     tasks.push(task0);
     tasks.push(task2);
     tasks.push(task3);
@@ -34,11 +33,7 @@ export class MyCaseWorkModel {
   }
 
   static getMyTasks(): Array<Task> {
-    const tasks: Array<Task> = [];
-
-    const task4 = new Task('1549 4765 3206 5586', 'Kili Muso', 'Protection', 'Taylor House', 'Review respondent evidence', 'Today', 3, 'today');
-    const task5 = new Task('1549 5366 1108 0150', 'Mankay Lit', 'Revocation', 'Taylor House', 'Review appellant case', '14 Dec', 1, 'future');
-    const task6 = new Task('1549 2345 7854 9786', 'Wina Palic', 'Human Rights', 'Taylor House', 'Review respondent evidence', '+4 days', 5, 'past');
+    const tasks: Array<Task> = [];    
 
     tasks.push(task4);
     tasks.push(task5);
