+++ src/main/routes/taskList.ts
@@ -1,18 +1,20 @@
 import * as express from 'express';
 import Debug from 'debug';
-import { createMyCaseWorkPage, claimTask, unClaimTask } from '../controllers/myCaseWork';
+import { createTaskListPage } from '../controllers/taskListController';
+import { claimTask } from '../controllers/claimTaskController';
+import { unClaimTask } from '../controllers/unClaimTaskController';
 import { filterTasksByLocations } from '../controllers/filterTasksByLocations';
 
-const debugMyCaseWork = Debug('app:route:myCaseWork');
 const debugFilter = Debug('app:route:filter');
 const debugClaimTask = Debug('app:route:claimTask');
 const debugUnclaimTask = Debug('app:route:unclaimTask');
+const debugTaskList = Debug('app:route:taskList');
 
 const router = express.Router();
 
-router.get('/my-case-work', (req, res) => {
-  debugMyCaseWork('my-case-work router...');
-  createMyCaseWorkPage(req, res);
+router.get('/task-list', (req, res) => {
+  debugTaskList('task-list router...');
+  createTaskListPage(req, res);
 });
 
 router.get('/filter', (req, res) => {
