+++ src/main/views/components/filter-by-location.njk
@@ -76,7 +76,7 @@
       text: "Done",
       preventDoubleClick: true,
       attributes: {
-        "id": "my-button"
+        "id": "filter-done-button"
         }
       }) }}
   </div>
