--- src/main/views/my-case-work.njk
@@ -1,102 +1,0 @@
-{% extends "template.njk" %}
-
-{% from "components/my-tasks.njk" import myTasks %}
-{% from "components/filter-by-location.njk" import filterByLocation %}
-{% from "govuk/components/details/macro.njk" import govukDetails %}
-
-{% block pageTitle %}GOV.UK - WA prototype frontend{% endblock %}
-
-{% block header %}
-  {{ govukHeader({
-    homepageUrl: "/",
-    containerClasses: "app-width-container",
-    serviceName: "Immigration & Asylum",
-    serviceUrl: "/",
-    navigation: [
-      {
-        href: "/",
-        text: "My cases"
-      },
-      {
-        href: "/my-case-work",
-        text: "My casework",
-        active: true
-      },
-      {
-        href: "#",
-        text: "All cases"
-      },
-      {
-        href: "#",
-        html: signOut()
-      }  
-    ]
-  }) }}
-{% endblock %}
-
-{% block content %}
-  
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-one-half">
-    <h1 class="govuk-heading-xl">My casework</h1>      
-  </div>
-</div>
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-one-half">
-    <h1 class="govuk-heading-l">My tasks</h1>      
-  </div>
-</div>
-
-{{ myTasks(tasks.myTasks, "myTasks") }}
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-one-half">
-      <h1 class="govuk-heading-l">My available tasks</h1>
-  </div>
-</div>
-
-<div class="govuk-grid-row">
-  <div id="ex1" class="govuk-grid-column-two-half">
-    {{ govukDetails({
-      summaryText: "Filter by location",
-      html: filterByLocation(tasks.addLocations, tasks.removeLocations),
-      classes: "filter-by-location-details"
-    }) }}    
-  </div>
-</div>
-
-{{ myTasks(tasks.myAvailableTasks, "myAvailableTasks") }}
-
-  
-{% endblock %}
-
-
-  {% block pageScripts %}
-  {% include "includes/sortable-table-scripts.njk" %}
-  {% include "includes/filter-by-location-scripts.njk" %}
-  {% include "includes/detail-actions-scripts.njk" %}
-  <script>
-    new MOJFrontend.SortableTable({
-      table: $('table')[0]
-    });
-    new MOJFrontend.SortableTable({
-      table: $('table')[1]
-    });
-
-    document.getElementById("my-button").onclick = () => {
-      const listItems = document.querySelectorAll('#ss_unimp_list li');
-      let query = "";
-      for (let step = 0; step < listItems.length; step++) {
-        query +="item" + step + '=' + listItems[step].innerText;
-        if (step < listItems.length - 1) {
-          query += "&";  
-        }
-       
-      }
-      location.href = "/filter?" + query;
-    }
-
-  </script>
- {% endblock %}
-
-
